refactor: Consolidate asset deletion dialog and fix error handling
  - Extract duplicate dialog logic into confirmDelete method
  - Only emit 'asset-deleted' event on successful deletion
  - Add missing i18n translation keys for delete messages
  - Remove unreachable code and improve null safety
